♻️ chore: Button 컴포넌트 사용법 변경 및 상수 변경
