+++ src/constant/size.js
@@ -1,11 +1,20 @@
+/**
+ * @enum {string}
+ */
 const SIZE = {
   MEDIUM: "md",
   SMALL: "sm",
   X_SMALL: "xs",
 };
 
+/**
+ * @enum {string}
+ */
 const PROFILE_SIZE = { ...SIZE };
 
+/**
+ * @enum {string}
+ */
 const BUTTON_SIZE = {
   ...SIZE,
   LARGE_34: "lg34",
